Change download location for Windows from scapy.net to hg repo
The ZIP file available at http://scapy.net/ has an additional header to
make it startable on Linux. This confuses the Windows internal unzipper
and causes warnings about corrupted archives. The files can be
extracted with 7-Zip or InfoUnzip, but this might be inconvenient
for a lot of users. So we now point to the Hg repo which carries
an unmodified ZIP.
* * *
Fixed duplicate reference in install.rst

The duplicate reference "latest version" gave a Sphinx warning.
* * *
Updated version number of Scapy documentation
* * *
Fixed typo in Windows installation doc