layers: full implementation of ipsec (esp & ah)
This patch adds a full implementation of the IPSec layer in
scapy/layers/ipsec.py. See the module's doc string for instructions.
In order to support encryption/decryption and authentication, the
pycrypto library is required. If not found, only the NULL/NULL
algorithms will be available.

(almost) complete test suite can be found in test/ipsec.uts and can
be launched with:

  ./bin/UTscapy -m scapy/layers/ipsec.py -t test/ipsec.uts -f text -F

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>

--HG--
branch : ipsec