layers/ipsec: fix padding problems with fuzzed packets
When generating packets with fuzz(), some fields are not initialized
completely. This causes problems in evaluating the amount of padding
necessary to align the encrypted data length to a multiple of the crypto
algorithm block size.

This patch forces the encrypted packet fields to be resolved to their
default values before calculating the pad length.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>