fragment6 now preserves custom values in the nh field.
fragment6 was not preserving the value of the nh (next header) field
when explicitly set. With something like
`IPv6()/IPv6ExtHdrFragment()/UDP()/'payload'` nh was preserved, but
with `IPv6()/IPv6ExtHdrFragment(nh=254)/'payload'` it was not. In the
latter case, the resulting fragments had a value of 59 (no next
header) in nh.