Workaround to ignore interfaces whose name got truncated by netstat
When interface names are longer than 7 characters netstat will truncate
them to the first 7 chars which leads to looking up an invalid interface
name.
We attempt to guess the name of the interface based on the output of
ifconfig -l. If there is only one candate we consider that otherwise we
ignore it.