Added a stop condition to PacketListField
Added a stop parameter to PacketListField. This parameter is a callable that would be evaluated after each packet is identified, and used as a condition to keep adding more packets to the list field. This enables the implementation of protocols who uses a "sentinel" or some property of an inner packet to mark the end of the packet list, in addition or instead of providing the count/length.