Allow to set promisc mode in sndrcv().
In the pcapdnet() backend it was not possible to run the
L{2,3}{pcap,dnet}Socket in promiscuous mode.  Add a promisc option
to sr(), sr1(), srp(), srp1(), srflood(), srpflood() so that the
user can choose the behavior.

Keep the default non-promiscuous as in most cases sndrcv() sniffs
an answer that is directed to the sender.