FlagValue: support multiple flags check at once
This PR makes it possible to check if several flags are all set, by
writting, for example, `if pkt[TCP].flags.SA:`.

It only works for single letter flags (e.g., will work for TCP().flags
but not for IP().flags), and only when getting a value.

Thanks @guedou for suggesting that.