Accept TCP RST packets with .ack == 0 as answers
Without this code, when an out of connection TCP packet receives a
TCP answer with flag RST and the ack set to 0, it is not seen as an
answer.

This comes without regression tests because I have been unable to
find a reliable machine without firewall on the Internet.

Simple test, if 192.168.0.1 has no firewall:

>>> sr(IP(dst="192.168.0.1")/TCP(sport=80, dport=12345, flags='SA'),
...    timeout=2)

This test will not terminate without this patch.