Only read version from git when in scapy
Scapy may be used as third party library by other projects, themselves
managed under git. This leads to surprising results such as this:

    $ git describe
    awesomeproject-v5.22-1-ga1316614c290
    $ virtualenv env
    ...
    $ . env/bin/activate
    (env)$ pip install scapy==2.3.3
    ...
    (env)$ cat env/lib/python2.7/site-packages/scapy/VERSION
    2.3.3
    (env)$ which scapy
    /home/.../env/bin/scapy
    (env)$ scapy
    ...
    Welcome to Scapy (awesomeproject-v5.22.dev1)
    >>>

Scapy's version is wrongly set to the current project's version.

When trying to determine scapy's version from git, make sure that the
code is executed from the root of a git repo. If not, read the version
from the scapy/VERSION file which has been generated when packaging
scapy source archive.

Fixes: 4f71027fcdcb ("enhance version management")
Signed-off-by: Robin Jarry <robin.jarry@6wind.com>