Fix EAP-MD5 Dissection
Implementation of EAP-MD5 was not following RFC3748 (which is
referencing RFC1994) properly.

Field value_size is determining only size of value_field.
It might have length different than 16B in EAP-MD5 request,
because in that case it is not used to transmit MD5 hash,
but random challenge value.
Size of optional_name(extra data) is determined as "rest of" EAP message
(len field of EAP).