layers/tls: fix shell start with old cryptography
Seen on Redhat-7 with distribution version of python-cryptography
(1.3.1):
  File "/usr/lib/python2.7/site-packages/scapy/layers/tls/crypto/ffdh.py", line 207, in <module>
    params = pn.parameters(default_backend())
AttributeError: 'DHParameterNumbers' object has no attribute 'parameters'

This is due to an API change in cryptography 1.7 which is the required version
for scapy. Althought, we should still be able to launch the shell.

Check if cryptography is valid before filling FFDH_GROUPS.

Signed-off-by: Thomas Faivre <thomas.faivre@6wind.com>