main: allow changing prompt when using ipython
IPython uses a specific system to change its prompt and does not rely on
sys.ps1

Modify the _prompt_changer function to handle IPython differently.

_prompt_changer is first called ealier than the actual launch of the
shell. Therefore, it is not enough to change the prompt using the
startup script. The embed function must be called with a configuration
object. This object is set to the same as embed would set it if not
provided, plus the prompt change.

Signed-off-by: Thomas Faivre <thomas.faivre@6wind.com>