config: allow a prompt prefix
Add a prompt_prefix configuration field that will be prepended to the
prompt.

This is interesting for applications that embed Scapy and want to prepend
a label to their prompt without having to worry about the type of shell
used (ipython for instance).

conf.prompt_prefix = 'Scapy '

Normal shell:

Scapy >>>

IPython:

Scapy In [1]:

Signed-off-by: Thomas Faivre <thomas.faivre@6wind.com>