tools/UTscapy: add option to break on failed campaign
When running several campaign, it can be interesting to run everything
despite the fact that a campaign has failed.
Add an option to keep the possibility to break at the first failed
campaign.

WARNING: this change the default current behavior. Travis configuration
files have been modified to keep current behavior.

Signed-off-by: Thomas Faivre <thomas.faivre@6wind.com>