Decodes the PCAP library version from ASCII bytes to a string
This is for Python 3 compatibility. Following the principle of decoding
when data enters the application, this seems to be the appropriate
place to do it.

Encoding bytes in Python 2 does not cause breakage, but decoding can.
For this reason, all string literals in the affected file are
treated as Unicode literals in Python 2 by the inclusion of
unicode_literals from __future__.