Linting and cleaning
Linted with flake8, ignoring a handful of items which seemed to be less
crucial F403,E266,F401,E402,F405

I was inspired to do this work because of all of the hanging whitespace.
When I saw the contributor docs are looking to move to PEP8, I wanted to
contribute some effort to clean up the codebase.

If this work is appreciated and merged, I'll move into the
subdirectories as I have time.