Manage received_frames per bpf socket.
When using sniff() and sndrcv() in multiple threads with different
bpf sockets, the global received_frames cache is disturbing.  Filters
or interfaces are not applied from the current sniff() invocation
as obsolete packets with different parameters may be already in the
received_frames array.  I would expect that the cache is part of
each bpf socket.  So it should be initialized in the constructor.