Calls tcpdump with `-nn` option, for compatibility with Red Hat's tcpdump.
This has no effect on upstream tcpdump.  _Without_ this change, the
regression tests fail on Red Hat distributions that carry this patch (all
versions released since 2001).

Background:

In 2001, Red Hat introduced a patch to their distribution of tcpdump which
splits the "don't resolve hostnames" and "don't resolve port names" options:

https://src.fedoraproject.org/rpms/tcpdump/c/d42d137

That patch causes `-n` to still resolve port names, so one gets results like
`192.0.2.1.ftp` instead of `192.0.2.1.21`.

tcpdump's mailing list archives go back to only 2002, and the earliest
attempt I could find to upstream this was in 2009, which I could find no
response to.

There was another attempt to upstream it in 2018 (as a "CentOS patch") that
has been explicitly _rejected_ by tcpdump:

https://github.com/the-tcpdump-group/tcpdump/pull/702