Option for treating response pending as answer
Until now the answers() function of GMLAN considered the
response-pending response as a proper answer. This led to manually
having to filter when receiving responses and eventually restart the
receiving until the proper answer (positive or negative response) is
received.

Similar to the implementation of UDS, response-pending is now not
considered an answer by default - this is the most common case.
Setting an option allows for treating these messages as a answer if
needed.