arch: fix unclosed socket warning
When forging a packet with no Ether source address and conf.iface holds
an interface that does not exist on the system, we get a pertinent
warning:

  WARNING: Could not get the source MAC: [Errno 19] No such device

However, it is followed by a resource warning:

  scapy/layers/l2.py:143: ResourceWarning: unclosed <socket.socket fd=8,
  family=AddressFamily.AF_INET, type=SocketKind.SOCK_STREAM, proto=0,
  laddr=('0.0.0.0', 0)>

Indeed, when the ioctl call fails, the socket is not closed. Make sure
to close it in all cases.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>