Refactor GTPv2 Information Elements a bit.
The main objective of this PR is to reduce duplication.  Currently,
Information Elements (IEs) type codes are duplicated in three places: in
the `IEType` dict, in defaults of `ietype` fields of each IE class, and
in `ietypecls` dict.  This change removes the numbers from the latter
two.

(And fixes one mismatch, `IE_MMBR.ietype` default was not correct.)

There's still duplication of IE names, but now it will raise `KeyError`
on module load time if there's mismatch.