volatile: add command method to be used in Packet.command()
When a VolatileValue subclass is used inside a Packet, the command()
method returns a string that get a Syntax Error when evaluated, e.g.:

> ICMPv6NIQueryNOOP(nonce=<RandBin>)

Add a command() method to the VolatileValue class so that it can be
handled properly. The _command_args() method is used to defined the
string used as argument for a given VolatileValue subclass.

Default values are not displayed.

Fixes #2828.

Signed-off-by: Thomas Faivre <thomas.faivre@6wind.com>