Use sock_fprog instead of bpf_program on Linux
sock_fprog is different from bpf_program in that its length field is
ushort instead of int. This results in a different layout on big-endian
machines.

Fix by defining sock_fprog and copying field values from bpf_program.

In theory sock_filter and bpf_insn could be affected too, but in
practice they are the same.

Signed-off-by: Ilya Leoshkevich <iii@linux.ibm.com>