WIP: use `ip route` to read IPv6 routes instead of /proc
The /proc/net interface has two disadvantages:

 1. It is very slow, taking about 2-3 ms per route (about 60 times
    slower than netlink).
 2. It doesn't support filtering by table, making it even slower *and*
    potentially confusing scapy when it sees multiple routes for the
    same destination.

For example, my system has ~60 routes in the 'main' table and ~120K
routes in the 'inet' table. It is useless for scapy to scan routes from
non-default tables because it does not process policy rules anyway; it
would be better if it only looked at the 'main' table.

With netlink it is possible to request a filtered dump (even though
getting the complete dump is already reasonably fast), but /proc/net
does not even have an indication of which table a route belongs to.)

Of course, it would be *much* better to use a native Python netlink
module instead of shelling out to `ip` (especially considering that the
-json option only became available last year),

But even the current approach shows quite an improvement:

 - /proc (all tables): 115690 routes in 304.03170132637024 seconds
 - iproute2 (all tables): 115651 routes in 5.262296199798584 seconds
 - iproute2 (main table): 32 routes in 0.07946968078613281 seconds

(The small differences in route count come mostly from BGP announcements
and withdrawals happening between dumps.)

Ref: secdev/scapy#2133