Relax a constraint in TLS Hello Extension handling.
The removed comment suggests that RFC 5246 forbids explicitly empty
extensions. However, RFC5246 describes the field, when present, to
be of the following type:

    Extension extensions<0..2^16-1>;

Which means that the variable length can be 0.

When scapy/tls speaks to some stacks (such as the Erlang one), we
can encounter a case where the ServerHello emitted by such stacks
will contain an extensions field with just 00 00, which can lead to
a handshake failure or strange behaviour from the automaton.

It stills seems to be a good idea not to emit the length field in
the build phase, when the actual extension list is empty.