Rename duplicate fields for contrib.scada.iec104
Fixes the last open checkbox of #2862 - the contrib.scada.iec104 part.

There were some places where multiple information elements (thus fields) were concatenated, e.g. the fields of the `IEC104_IE_SIQ` and the `IEC104_IE_CP56TIME2A` classes. The first one is a SIQ (single point information with quality descriptor) and the second one is a CP56Time2a (basically a timestamp), so the concatenated result could be called a `single-point information with quality descriptor and with time tag cp56time2a`. Both of which contain a field called `iv` leading to the (in this case very correct) warning as this results in multiple fields in a packet to have the same name.

In short: There are multiple classes / data types in the standard that include a time tag cp56time2a, thus leading to this warning in many places.

So there were two possibilites: Changing all of the classes that use `iv` as their field name or simply rename this field in the cp56time2a time tag field. I chose the latter as this has a smaller impact on existing code. Obviously, code addressing exactly that field will now have to be adjusted accordingly. However, I'm not sure how and if the old code worked anyways as they had exactly the name. So I guess, this change is not too destructive.

The same applies for the `reserved` field in the `IEC104_IE_QDP` class that gets concatenated to a number of classes - though, in this case probably noone ever used / addressed the `reserved` field anyways.

This MR gets rid of all the syntax warnings (and my code using this part of scapy continues to work - however, I don't use any of the types that make use of these fields). It tries to introduce as few breaking changes as possible (after a thorough investigation, I found that only two of the field names needed to be changed, even though there were dozens of warnings).

Would be great if this MR would be merged (or if there is some feedback for changes first, I'm happy to hear about it!), because currently using the `scada.iec104` protocol in scapy spams the output with lots of syntax warnings and this reduces them to zero.

---

As a workaround in the meantime, I currently suppress the warnings until this is fixed:

```python
# ignore SyntaxWarnings from iec104 modules as they spam the complete log and it's not under our control anyways
import warnings
warnings.simplefilter('ignore', SyntaxWarning)
import scapy.contrib.scada.iec104 as iec104
warnings.simplefilter('default', SyntaxWarning)  # reset to default behaviour to not miss possible future helpful SyntaxWarnings
```