Implement coexistence of IKEv1 and IKEv2
Previously, it was not possible to dissect IKEv1 and IKEv2 packets
at the same time, loading `scapy.contrib.ikev2` manually broke
(replaced) the ISAKMP bindings.

This commit fixes the problem by adding a dispatch hook to the
IKEv2 layer which switches between ISAKMP and IKEv2 based on the
protocol version.

It also removes port 4500(ipsec_nat_t) as lower binding from IKEv2
to UDP, which breaks the binding for UDP-encapsulated ESP.
One of the unit tests had to be adjusted to match this change.