Implement UDP encapsulation of IPsec ESP packets [RFC 3948]
UDP encapsulation of IPsec ESP packets is used for NAT-traversal
on port 4500(ipsec-nat-t). Previously, the dissector assumed that
all traffic on port 4500 was UDP encapsulated ESP only.

This commit introduces a new UDP_ENCAP Packet which selects the
correct packet format for the payload as specified in the RFC.

Note that the implementation works for both IKEv1 (ISAKMP) and IKEv2,
but only after the 'scapy.contrib.ikev2' module is loaded.
Creating default NAT-T bindings for ISAKMP only and rebinding them
to IKEv2 when its loaded was just too ugly and painful. NAT-T support
by default for all IKE versions will be available as soon as IKEv2
is promoted to official scapy layer.