Implement UDP encapsulation of IPsec ESP packets
UDP encapsulation of IPsec ESP packets is defined in RFC3948
and used for NAT-Traversal on port 4500(ipsec-nat-t)

Previously, the dissector assumed that all traffic on port 4500
was UDP encapsulated ESP only.

This commit introduces a new UDP_ENCAP Packet which selects the
correct packet format for the payload (ESP vs. NON_ESP/ISAKMP
vs. NAT-KEEPALIVE) as specified in RFC3948, section 2.