LLDP: fix support of non-IPv4 network addresses in chassis/port IDs (#3773)
LLDPDUChassisID and LLDPDUPortID have a subtype "network
address" (0x05 and 0x04 respectively) for the "id" field that
interprets it as having the first octect specifying an address family,
and the other octects being interpreted in the specified address
family. We were supporting the specific parsing of IPv4 in this field,
and that was working fine.

The other possible address families were not working because:
- we were not checking the family field when choosing IPField as the
  class for the id and
- we were not subtracting the the family field from the length of the
  id in length_from when using StrLenField as the type.

These issues prevented the parsing and crafting of packets where the
address family is arbitrary or simply different from IPv4.

Besides fixing and adding tests, this commit also adds support for
IPv6 as a specific parser for the network address field, when the
address family is 0x02.