IKEv2: test dissection of some decrypted IKEv2 payloads
This commit adds two testcases to the advanced test, which verify
dissecting and building for the two decrypted IKE_AUTH messages
of the pcap file.

 * IKEv2 keyexchange with NAT-traversal

   - IKE_SA_INIT request
   - IKE_SA_INIT response
   - IKE_AUTH request
   - IKE_AUTH response
   - IKE_AUTH request, decrypted    (new)
   - IKE_AUTH response, decrypted   (new)

Effectively, those two testcases combine unit tests for the following
payloads transmitted in the Encrypted payload of the IKE_AUTH exchange:

    AUTH, CERTREQ, CERT_CRT, CP, IDi, IDr, Nonce, Notify, Proposal,
    SA, TSi, TSr, Transform, VendorID

Also add a forgotten reference to the origin of the capture file: it is
taken from 'Example 2: Dissection of encrypted (and UDP-encapsulated)
IKEv2 and ESP messages' on the Wireshark [SampleCaptures] Wiki page.

Note: The tarfile not only contains the pcap file but also the secrets
which enable Wireshark (v3.6+) to decrypt the IKE and ESP traffic.
The raw frame data was crafted manually using Wireshark for decrypting
the encrypted payload and Scapy for gluing the layers together.

[SampleCaptures]: https://gitlab.com/wireshark/wireshark/-/wikis/SampleCaptures