IKEv2: remove the unnecessary 'payload_' infixes
The IKE messages tend to have a lot of payloads and those unnecessary
eight characters per payload add a lot of redundancy and size to the
output of the dissected packets. Removing them improves readability of
the output and makes typing less tedious.

The removal was automated using the following find+sed script:

    find * -type f -exec sed -i \
        -e 's/IKEv2_payload_type/IKEv2_payloax_type/g' \
        -e 's/IKEv2_payload_/IKEv2_/g' \
        -e 's/IKEv2_payload/IKEv2_Payload/g' \
        -e 's/IKEv2_payloax/IKEv2_payload/g' {} +