IKEv2: reorganize IKEv2{Attribute,Payload,Exchange}Types dicts
* IKEv2AttributeTypes

Swap names and numbers in the IKEv2AttributeTypes dictionary and its
nested algorithm dictionaries to get rid of the awkward code to invert
the entries. With this change, the transform types and algorithms can
now be extracted using a simple dictionary comprehension.

* IKEv2PayloadTypes and IKEv2ExchangeTypes

Instead of arrays, use dictionaries like everywhere else in the module.
Also change their names to make them consistent with the names of the
other dictionaries.