IKEv2: use IANA abbreviations for most of the payloads
Replace the names of the 'Encrypted', 'Encrypted_Fragement, and
'VendorID' payload with their official IANA abbreviations 'SK',
'SKF, and 'V', to be more consistent with the other payload names.

The only remaining exceptions are the 'Nonce' and 'Notify' payloads
('N' resp. 'Ni'/'Nr'), because they would end up with the same
abbreviation (the dissector does not distinguish between the
initiator's and responder's nonce).