IKEv2: improve dissection of the CERT and CERTREQ payload
* CERT payload  [RFC 7296, section 3.6]

Previously, the IKEv2_CERT class used a dispatch hook to define three
different classes IKEv2_CERT_CRT, IKEv2_CERT_CRL, and IKEv2_CERT_STR.

This commit removes the three subclasses and uses a MultipleTypeField
for the `cert_data` instead.

* CERTREQ payload  [RFC 7296, section 3.7]

The field description was apparently copied from the vendor id payload
and never updated. This commit adds the correct members, however the
size and count of the SPIs are not updated automatically yet.