Fix error handling for pcap_create() and pcap_activate().
pcap_create() is not guaranteed to succeed; check whether it returns a
null POINTER and, if so, report the error.

pcap_activate() can return

1) 0, which means it succeeded with no warnings;
2) a netative number, which means it failed, and the number gives an
indication of the type of failure
3) a positive number, which means it succeeded, but with a warning, and
the number gives an indication of the type of warning.

For now, we don't bother with warnings, so we check only for a negative
return value.

If we get a negative return value, then:

if the return value is PCAP_ERROR, report an error with the result of
pcap_geterr() as the error message;

if the return value is PCAP_ERROR_NO_SUCH_DEVICE, report an error with
the interface name, the result of pcap_statustostr() and the result of
pcap_geterr() all in the error message

if the return value is PCAP_ERROR_PERM_DENIED, and the result of
pcap_geterr() isn't an empty string, report an error with the interface
name, the result of pcap_statustostr() and the result of pcap_geterr()
all in the error message;

otherwise, report an error with the interface name and the result of
pcap_statustostr() in the error message.

(This is modeled after what tcpdump does.)