IKEv2: replace Str*Fields with XStr*Fields
All the fields contain binary data, so it makes the output
much more readable to show their value as hexadecimal string.
It is also more consistent with the ipsec module, see for
example the spi and data fields of the ESP packet.

Additionally, unify naming and formatting of lambda functions and
rename the 'load' field of the Nonce, Notify and KE payloads to
'nonce', 'notify' resp. 'ke'.