Fix bug #191
BluetoothUserSocket did not close the first socket returned
by the first call to socket(), leaving this socket alive.
OS returns "Device or resource busy" because of this when
one try to instanciate a new BluetoothUserSocket.

Closing this first socket in BluetoothUserSocket's `close()`
method free all the file descriptors and solved this issue.