Fix the "Classless Static Route" parser
FieldListField expects its fields to either fully consume bytes or throw
exceptions if those bytes can't be consumed. The "Classless Static
Route" field didn't do that when it received invalid subnet mask widths
and it led to a loop where FieldListField tried to parse the same
sequence of bytes and append the same value to the list over and over
again until scapy got interrupted manually or got killed by the OOM killer.

This patch addresses that by throwing exceptions when classless static
routes can't be consumed. It basically relies on inet_ntoa rejecting
anything other than 4 bytes. The test suite is updated to cover the
corner cases. Other than that the DHCP parser was fuzzed for about an
hour and nothing popped up.

It's a follow-up to f5b718881441637e9f8cab75515981b32247f707