IKEv2: refactor the implementation (#3795)
* IKEv2: remove the unnecessary 'payload_' infixes

The IKE messages tend to have a lot of payloads and those unnecessary
eight characters per payload add a lot of redundancy and size to the
output of the dissected packets. Removing them improves readability of
the output and makes typing less tedious.

The removal was automated using the following find+sed script:

    find * -type f -exec sed -i \
        -e 's/IKEv2_payload_type/IKEv2_payloax_type/g' \
        -e 's/IKEv2_payload_/IKEv2_/g' \
        -e 's/IKEv2_payload/IKEv2_Payload/g' \
        -e 's/IKEv2_payloax/IKEv2_payload/g' {} +

* IKEv2: reorganize IKEv2{Attribute,Payload,Exchange}Types dicts

* IKEv2AttributeTypes

Swap names and numbers in the IKEv2AttributeTypes dictionary and its
nested algorithm dictionaries to get rid of the awkward code to invert
the entries. With this change, the transform types and algorithms can
now be extracted using a simple dictionary comprehension.

* IKEv2PayloadTypes and IKEv2ExchangeTypes

Instead of arrays, use dictionaries like everywhere else in the module.
Also change their names to make them consistent with the names of the
other dictionaries.

* IKEv2: bind all layers together correctly

The previous bindings were incomplete: the payloads were only bound
to the main IKEv2 layer, but not to the other payloads. For example,
the lower binding between KE and Nonce payload is missing here:

    >>> IKEv2() / IKEv2_KE() / IKEv2_Nonce()
    <IKEv2  next_payload=KE |<IKEv2_KE  |<IKEv2_Nonce  |>>>

This commit removes the manually implemented bindings and replaces them
with bind_layers() calls:

For every entry (np, name) in the IKEv2PayloadTypes dictionary it is
now expected that there exists a corresponding 'IKEv2_<name>' class,
which is derived from IKEv2_Packet. The layers get bound as follows:

    bind_layers(IKEv2_Packet, IKEv2_<name>, next_payload=np)

* IKEv2: improve dissection of the CERT and CERTREQ payload

* CERT payload  [RFC 7296, section 3.6]

Previously, the IKEv2_CERT class used a dispatch hook to define three
different classes IKEv2_CERT_CRT, IKEv2_CERT_CRL, and IKEv2_CERT_STR.

This commit removes the three subclasses and uses a MultipleTypeField
for the `cert_data` instead.

* CERTREQ payload  [RFC 7296, section 3.7]

The field description was apparently copied from the vendor id payload
and never updated. This commit adds the correct members, however the
size and count of the SPIs are not updated automatically yet.

* IKEv2: replace Str*Fields with XStr*Fields

All the fields contain binary data, so it makes the output
much more readable to show their value as hexadecimal string.
It is also more consistent with the ipsec module, see for
example the spi and data fields of the ESP packet.

Additionally, unify naming and formatting of lambda functions and
rename the 'load' field of the Nonce, Notify and KE payloads to
'nonce', 'notify' resp. 'ke'.